Add a list-styli command
Only way to list the styli right now is to parse the control-center
cache file [1] where the serial number is used as section header. That
file would also include the tool ID but to parse that we'd need libwacom
bindings for Python.

(we could dig around in dconf but libdconf doesn't have GIR bindings so
it's in the "too hard" basket right now)